requirements.txt
python-telegram-bot==22.0
paypalrestsdk==1.13.3
python-dotenv
Flask
requests
